// frontend/src/oas/models/Train.ts
export interface Train {
    /**
     * Train (列車) の説明
     * @type {string}
     * @memberof Train
     */
    description: string;
    /**
     * 列車番号
     * @type {string}
     * @memberof Train
     */
    trainNumber: string;
    /**
     * 進行方向 (0~1: 下り, -1: 上り)
     * @type {number}
     * @memberof Train
     */
    direction: number;
    /**
     * 仕業の初日からの経過日数 (0で初日/日勤、1で明け)
     * @type {number}
     * @memberof Train
     */
    dayCount: number;
    /**
     * TrainのID (UUID)
     * @type {string}
     * @memberof Train
     */
    readonly trainsId?: string;
    /**
     * WorkのID (UUID)
     * @type {string}
     * @memberof Train
     */
    readonly worksId?: string;
    /**
     * 作成日時
     * @type {Date}
     * @memberof Train
     */
    readonly createdAt?: Date;
    /**
     * 最高速度 (km/h)
     * @type {string}
     * @memberof Train
     */
    maxSpeed?: string;
    /**
     * 速度種別
     * @type {string}
     * @memberof Train
     */
    speedType?: string;
    /**
     * けん引定数
     * @type {string}
     * @memberof Train
     */
    nominalTractiveCapacity?: string;
    /**
     * 編成両数 (0以下で非表示)
     * @type {number}
     * @memberof Train
     */
    carCount?: number;
    /**
     * 行先
     * @type {string}
     * @memberof Train
     */
    destination?: string;
    /**
     * 乗車前の注意事項 (「乗継」など)
     * @type {string}
     * @memberof Train
     */
    beginRemarks?: string;
    /**
     * 降車後の注意事項 (「乗継」など)
     * @type {string}
     * @memberof Train
     */
    afterRemarks?: string;
    /**
     * 注意事項
     * @type {string}
     * @memberof Train
     */
    remarks?: string;
    /**
     * 発前
     * @type {string}
     * @memberof Train
     */
    beforeDeparture?: string;
    /**
     * 着後
     * @type {string}
     * @memberof Train
     */
    afterArrive?: string;
    /**
     * 列車に関する情報
     * @type {string}
     * @memberof Train
     */
    trainInfo?: string;
    /**
     * 添乗での移動かどうか
     * @type {boolean}
     * @memberof Train
     */
    isRideOnMoving?: boolean;
}